+++ packages/orq-rc/src/models/operations/runagent.ts
@@ -4031,7 +4031,7 @@ export const Tools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   Tools
 > = z.object({
-  id: z.string().default("01KJCGPR095T35R6V6147X05GF"),
+  id: z.string().default("01KJCMEG0ZVHFVMC1YVVZWCRE1"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),